Code refactor

+++ ASE.ScienceDirectExtractor/ASE.ScienceDirectExtractor.Data/AuthorRelation.cs
@@ -7,7 +7,7 @@ namespace ASE.ScienceDirectExtractor.Data
         public long Id { get; set; }
         public long Author1Id { get; set; }
         public long Author2Id { get; set; }
-        public List<long> Years { get; set; }
+        public List<int> Years { get; set; }
 		public bool SameAffiliationCity { get; set; }
 
 		public AuthorRelation(long id, long author1Id, long author2Id, bool sameAffiliationCity)
@@ -17,7 +17,7 @@ namespace ASE.ScienceDirectExtractor.Data
 	        Author2Id = author2Id;
 			SameAffiliationCity = sameAffiliationCity;
 
-            Years = new List<long>();
+            Years = new List<int>();
         }
     }
 }
